@@ -3,19 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "node server.js",
+    "dev": "node --max-old-space-size=4076 server.js",
     "build": "next build",
     "start": "cross-env NODE_ENV=production node server.js"
   },
   "dependencies": {
-    "@material-ui/core": "^4.11.2",
+    "@date-io/moment": "^1.3.13",
+    "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.57",
+    "@material-ui/pickers": "^3.3.10",
     "axios": "^0.21.0",
     "body-parser": "^1.19.0",
     "cors": "^2.8.5",
     "cross-env": "^7.0.2",
-    "dotenv": "^8.2.0",
+    "dotenv": "^8.6.0",
     "dotenv-webpack": "^6.0.0",
     "express": "^4.17.1",
     "fecha": "^4.2.0",
@@ -26,18 +28,20 @@
     "googleapis": "^67.0.0",
     "html-to-text": "^6.0.0",
     "image-upload-react": "^1.4.9",
+    "jsonwebtoken": "^8.5.1",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.32",
     "multer": "^1.4.2",
-    "next": "9.5.5",
+    "next": "^10.0.9",
     "next-images": "^1.6.2",
     "node-schedule": "^2.0.0",
     "nodemailer": "^6.4.17",
     "path": "^0.12.7",
     "pg": "^8.4.2",
     "postcss-preset-env": "^6.7.0",
-    "react": "^17.0.1",
+    "react": "^17.0.2",
     "react-contenteditable": "^3.3.5",
+    "react-countup": "^4.3.3",
     "react-dom": "17.0.1",
     "react-easy-crop": "^3.3.1",
     "react-html-parser": "^2.0.2",
