@@ -27,6 +27,12 @@ const LoginForm = (props) => {
                     break;
             }
         })
+        if (context.profile && !context.profile.isVerified){
+            props.showVerify();
+        }
+        if (context.profile && context.profile.isVerified){
+            props.showAccount();
+        }
     }
      
     const SigninSchema = Yup.object().shape({
@@ -41,14 +47,14 @@ const LoginForm = (props) => {
         <div className="grid grid-cols-2 mb-2">
             <div className="grid col-span-1">
                 <div className="flex flex-col">
-                    <p>Sign in / Sign up</p>
-                    <p className="my-4">To attend or host events, please create a Schefs account</p>
+                    <p>Sign in / Sign up</p> 
+                    <p className="my-4">To attend or host events, please create a Schefs account.</p>
                     <p>You wil need a valid university email to use Schefs.</p>
                 </div>
             </div>
             <div className="grid col-span-1">
                 <div className="flex flex-col">
-                    <img className="justify-self-center self-center h-40" src={sampleCard}></img>
+                    <img className="justify-self-center self-center h-48" src={sampleCard}></img>
                 </div>
             </div>
         </div>
@@ -58,28 +64,35 @@ const LoginForm = (props) => {
             //validator={() => ({})}
             validationSchema={SigninSchema}
         >
-            {({isValid, dirty}) => (
+            {({isValid, dirty, errors}) => (
             <Form>
                 <div className="grid my-4">
                     <p className="text-red-500 text-sm">{error[1]}</p>
                     <ErrorMessage render={msg => <p className="text-red-500 text-sm">{msg}</p>} name="email"></ErrorMessage>
                     <ErrorMessage render={msg => <p className="text-red-500 text-sm">{msg}</p>} name="password"></ErrorMessage>
-                        <div className="flex flex-col">          
+                        <div className="flex flex-col">    
                                 <Field placeholder="School Email" className={"border-2 border-solid rounded-full focus:outline-none my-2 px-4 py-1 " + error[2]} name="email"></Field>
                                 <Field type="password" placeholder="Password" className={"border-2 border-solid rounded-full focus:outline-none my-2 px-4 mb-3 py-1 " + error[3]} name="password"></Field>
                             <div className="mx-auto">
                                 <button disabled={!isValid || !dirty} type="submit" className={"flex px-16 mt-4 mb-2 py-0 justify-center items-center bg-transparent focus:outline-none text-black border sm:border-2 border-black rounded-full " + (!isValid || !dirty ? "cursor-not-allowed": "cursor-pointer hover:bg-black hover:text-white") }>LOG IN</button>
                             </div>
                         </div>
+                        <a id="forgotPword" onClick={props.resetFunction} className="underline text-center cursor-pointer justify-self-center">Forgot Your Password?</a>
                 </div>
+                <footer className={"justify-between flex " + 
+                    (Object.keys(errors).length == 2 
+                        ? "mt-16" 
+                        : Object.keys(errors).length == 1 
+                            ? "mt-20 pt-2"
+                            : error[1] 
+                                ? "mt-16"
+                                : "mt-24")}>
+                    <p>Don't have an account?</p>
+                    <WhitePillButton handleClick={props.function} text="SIGN UP" padding="flex px-6"/>
+                </footer>
             </Form>
             )}
         </Formik>
-        <a onClick={props.resetFunction} className="underline text-center cursor-pointer justify-self-center">Forgot Your Password?</a>
-        <footer className="my-2 mt-6 justify-between flex">
-            <p>Don't have an account?</p>
-            <WhitePillButton handleClick={props.function} text="SIGN UP" padding="flex px-6"/>
-        </footer>
     </div>
     );
 }
