@@ -2,16 +2,18 @@ import React from "react";
 import BlackPillButton from "../Buttons/bpillbutton"
 
 const Banner = (props) => {
-    let hostButtonProps = {
+    const hostButtonProps = {
         text: props.buttonText,
         size: "sm",
-        link: "/about"
+        link: "/eventbuilder"
     }
+
     return (
         <div>
-            <div className="z-40 h-8 md:flex justify-center md:justify-between md:px-8 md:fixed md:w-full items-center md:bg-black">
-                <p className="hidden md:block text-white text-xs">{props.left}</p>
-                <div className="hidden md:block">
+            <div className="z-40 h-10 md:flex justify-center md:justify-between md:fixed md:w-full items-center md:bg-black">
+                <p className="text-white pl-2 ml-6 md:ml-12 xl:ml-24 hidden lg:block">{props.left}</p>
+                <p className="text-white pl-2 ml-6 md:ml-12 xl:ml-24 lg:hidden md:block">{props.left_small}</p>
+                <div className="hidden md:block pr-2">
                     <BlackPillButton {...hostButtonProps} />
                 </div>
             </div>    
