@@ -10,12 +10,12 @@ const Footer = (props) => {
     }
     return (
         <div className="h-6">
-            <footer className="inset-x-0 fixed bottom-0 flex items-center px-8 justify-center bg-white h-8 md:justify-between md:text-xs md:h-6 md:bg-gray-300">
-                <div className="block md:hidden">
+            <footer id="footer" className="inset-x-0 fixed bottom-0 flex items-center h-8 md:h-10 px-8 md:px-12 xl:px-24">
+                <div id="footerBreakpointBtn">
                     <PillButton {...hostButton} />
                 </div>
-                <div className="hidden md:block"> {props.left}</div>
-                <div className="hidden md:block">{props.right}
+                <div id="footerBreakpointText"> {props.left}</div>
+                <div id="footerBreakpointText">{props.right}
                     <a className="underline" href="/about">{props.linkText}</a>
                 </div>
             </footer>
