@@ -3,7 +3,7 @@ import logo from "../../assets/logo.png"
 
 // give scrollShadow: true for shadow on scroll
 const NavBar = (props) => {
-    if (props.scrollShadow){
+    if (props.scrollShadow) {
         const [isScroll, setScroll] = useState(0);
         let last_known_scroll_position = 0;
     
@@ -16,37 +16,17 @@ const NavBar = (props) => {
     
             return () => window.removeEventListener("scroll", onScroll)
         }, [])
+
         return (
-            <div>
-                <nav style={{zIndex:5}} className={"bg-white fixed top-0 md:top-auto w-full  " + (isScroll !== 0 ? 'shadow-lg' : '')}>
-                    <div className="sm:flex sm:justify-start sm:mx-8 sm:px-0 my-4 mx-4 px-4 py-1">
-                        <div className="flex items-center">
-                            <a href="/index">
-                                <img className="h-6" src={logo}></img>
-                            </a>
-                            <div className={"sm:flex text-black "}>
-                            <a className="text-sm hover:text-gray-600 duration-300 ml-4" href="#">Host</a>
-                            <a className="text-sm hover:text-gray-600 duration-300 ml-4" href="/about">About</a>
-                            <a className="hidden sm:inline text-sm hover:text-gray-600 duration-300 ml-4" href="/openmindarchive">Open Mind Archive</a>
-                            <a className="sm:hidden text-sm hover:text-gray-600 duration-300 ml-4" href="/openmindarchive">OMA</a>
-                            <a className="hidden sm:inline text-sm hover:text-gray-600 duration-300 ml-4" href="/about">Ambassador Program</a></div>
-                        </div>
-                    </div>
-                </nav>
-            <div className="h-10 md:h-20"></div>
-          </div>
-        )
-    }
-    return (
-        <div>
-            <nav className={"bg-white fixed w-full top-0 md:top-auto"}>
-                <div className="sm:flex sm:justify-start sm:mx-8 sm:px-0 my-4 mx-4 px-4 py-1">
+          <>
+            <nav className={"bg-white fixed z-10 top-0 md:top-auto w-full " + (isScroll !== 0 ? 'shadow-lg' : '')}>
+                <div className="relative sm:flex sm:justify-start pl-2 ml-6 md:ml-12 xl:ml-24 mt-6 mb-4 py-1">
                     <div className="flex items-center">
                         <a href="/index">
-                            <img className="h-6" src={logo}></img>
+                            <img src={logo} style={{height: '1.75rem'}}></img>
                         </a>
                         <div className={"sm:flex text-black "}>
-                        <a className="text-base hover:text-gray-600 duration-300 ml-4" href="/eventbuilder">Host</a>
+                        <a className="text-base hover:text-gray-600 duration-300 ml-4 md:ml-8" href="/eventbuilder">Host</a>
                         <a className="text-base hover:text-gray-600 duration-300 ml-4" href="/about">About</a>
                         <a className="hidden sm:inline text-base hover:text-gray-600 duration-300 ml-4" href="/openmindarchive">Open Mind Archive</a>
                         <a className="sm:hidden text-base hover:text-gray-600 duration-300 ml-4" href="/openmindarchive">OMA</a>
@@ -54,8 +34,30 @@ const NavBar = (props) => {
                     </div>
                 </div>
             </nav>
-        <div className="h-8 md:h-16"></div>
-      </div>
+            <div className="h-12 md:h-20 md:mb-6"></div>
+          </>
+        )
+    }
+
+    return (
+      <>
+        <nav className="bg-white z-10 fixed w-full top-0 md:top-auto">
+            <div className="relative z-50 sm:flex sm:justify-start pl-2 ml-6 md:ml-12 xl:ml-24 mt-6 mb-4 py-1">
+                <div className="flex items-center">
+                    <a href="/index">
+                        <img src={logo} style={{height: '1.75rem'}}></img>
+                    </a>
+                    <div className={"sm:flex text-black "}>
+                    <a className="text-base hover:text-gray-600 duration-300 ml-4 md:ml-8" href="/eventbuilder">Host</a>
+                    <a className="text-base hover:text-gray-600 duration-300 ml-4" href="/about">About</a>
+                    <a className="hidden sm:inline text-base hover:text-gray-600 duration-300 ml-4" href="/openmindarchive">Open Mind Archive</a>
+                    <a className="sm:hidden text-base hover:text-gray-600 duration-300 ml-4" href="/openmindarchive">OMA</a>
+                    <a className="hidden sm:inline text-base hover:text-gray-600 duration-300 ml-4" href="/about">Ambassador Program</a></div>
+                </div>
+            </div>
+        </nav>
+        <div className="h-12 md:h-20 md:mb-6"></div>
+      </>
     )
 };
 
