@@ -7,12 +7,26 @@ import CardContent from './usercardcontent'
 import SignUpForm from '../Auth/signupform';
 import LoginForm from '../Auth/loginform';
 import PasswordReset from '../Auth/passwordreset';
+import VerifyEmail from '../Auth/verifyemail';
 import Context from '../Context/context';
 
 // props.right = boolean side of screen
 const Card = (props) => {
     const context = useContext(Context);
-    const [cardInterior, setCardInterior] = useState("login");
+    const [cardInterior, setCardInterior] = useState("");
+    const [email, setEmail] = useState("");
+
+    useEffect(() => {
+        if (context.profile && !context.profile.isVerified){
+            setCardInterior("verify")
+        }
+        if (!context.profile) {
+            setCardInterior("login")
+        }
+        if (context.profile && context.profile.isVerified){
+            setCardInterior("account")
+        }
+    }, [context.profile])
 
     const handleCloseCard = () => context.handleCloseCard(!props.right, props.right);
 
@@ -24,6 +38,10 @@ const Card = (props) => {
             setCardInterior("login")
         }
     }
+    const verify = (e) => {
+        setCardInterior("verify");
+        setEmail(e);
+    }
     return (
         <CSSTransition
           in={props.right ? context.rCardIsOpen : context.lCardIsOpen}
@@ -32,7 +50,7 @@ const Card = (props) => {
           classNames={props.right ? 'rcard' : 'lcard'}
           unmountOnExit
         >
-          <div id="cardSize" className={'fixed m-2 border sm:border-2 border-black rounded-xl md:mt-10 top-0 bg-white justify-center z-20 ' + (props.right ? 'right-0' : 'left-0')}>
+          <div id="cardSize" className={'fixed m-2 border sm:border-2 border-black rounded-3xl md:mt-12 top-0 bg-white justify-center z-20 ' + (props.right ? 'right-0' : 'left-0')}>
             <div className="flex justify-end">
               <button onClick={handleCloseCard} className="focus:outline-none p-2">
                 <HighlightOff/>
@@ -40,15 +58,17 @@ const Card = (props) => {
             </div>
 
 	        {props.right
-		      ? context.profile
+		      ? cardInterior === "account" && context.profile
                   ? <CardContent profile={context.profile} />
-                  : cardInterior === "login"
-                      ? <LoginForm function={toggle} resetFunction={() => setCardInterior("password")}/> 
-                      : cardInterior === "signup" 
-                          ? <SignUpForm function={toggle} /> 
-                          : cardInterior === "password"
-                              ? <PasswordReset function={() => setCardInterior("login")}/> 
-                              : null
+                    : cardInterior === "login"
+                        ? <LoginForm function={toggle} resetFunction={() => setCardInterior("password")} showVerify={() => setCardInterior("verify")} showAccount={() => {setCardInterior("account")}}/>
+                        : cardInterior === "verify" && context.profile
+                            ? <VerifyEmail email={email} function={() => setCardInterior("login")}/>
+                            : cardInterior === "signup" 
+                                ? <SignUpForm function={toggle} showVerify={verify}/> 
+                                : cardInterior === "password"
+                                    ? <PasswordReset function={() => setCardInterior("login")}/> 
+                                    : null
               : context.leftProfile
                   ? <CardContent profile={context.leftProfile} />
                   : null
