@@ -8,10 +8,11 @@ const CardButton = () => {
     const [inHover, setHover] = useState("#000000");
     return (
         <>
-        <button onClick={() => context.handleToggleCard(false, true)} className="focus:outline-none fixed z-10 flex pb-2 justify-center items-center bg-gray-200 shadow-lg rounded-xl right-0 top-0 w-16 h-20 mt-2 mr-2 sm:w-20 sm:h-24 sm:mt-10 hover:bg-gray-700" onMouseEnter={() => setHover("#FFFFFF")} onMouseLeave={() => setHover("#000000")}>
+        <button id="cardBtn" onClick={() => context.handleToggleCard(false, true)} className="focus:outline-none fixed z-10 flex pb-2 justify-center items-center rounded-xl right-0 top-0 mt-2 mr-2 w-20 h-24 md:w-24 md:h-32 xl:w-32 xl:h-40 sm:mt-12" onMouseEnter={() => setHover("#FFFFFF")} onMouseLeave={() => setHover("#000000")} style={{boxShadow: '2px 2px 4px rgba(0, 0, 0, 0.25)'}}> 
           <Person style={{fontSize: 30}} htmlColor={inHover}/>
         </button>
         </>
     )
 }
 export default CardButton
+//bg-gray-200 hover:bg-gray-800
