@@ -14,7 +14,7 @@ const GreyOut = () => {
           unmountOnExit
         >
           <div onClick={() => context.handleCloseCard(true, true)} className="fixed inset-0 z-10">
-            <div className="absolute inset-0 bg-gray-700 bg-opacity-75"></div>
+            <div className="absolute inset-0 bg-black" style={{opacity: .4}}></div>
           </div>
         </CSSTransition>
     )
