@@ -1,20 +1,22 @@
 const moment = require('moment-timezone');
 const schedule = require('node-schedule');
 
-const { verifyFirebaseIdToken } = require('../middleware/auth');
+const { verifyFirebaseIdToken, verifyIsAdmin } = require('../middleware/auth');
 
 const express = require('express');
 const pool = require('../utils/db');
 const queries = require('../utils/queries/events');
 const thumbnail_queries = require('../utils/queries/thumbnails');
 const emails = require('../utils/emails');
+const gcal = require('../utils/gcalendar');
+const zoom = require('../utils/zoom');
 
 const router = express.Router();
+const TIMEZONE = 'America/New_York';
 
-// TODO: event approval/denial
 /*
  * GET /api/events/countTickets
- * Get number of all tickets reserved from .
+ * Get number of all tickets reserved in given time span.
  *
  * Request Parameters:
  *  query:
@@ -28,28 +30,45 @@ const router = express.Router();
  *  500: other postgres error
  */
 router.get('/countTickets', (req, res) => {
-    pool.query(queries.getAllReservedTicketsCount, (q_err, q_res) => {
-        if (q_err)
+    pool.query(queries.getAllReservedTicketsCount, [ req.query.date_from, req.query.date_to ], (q_err, q_res) => {
+        if (q_err){
             res.status(500).json({ err: 'PSQL Error: ' + q_err.message });
+            console.log(q_err.message);
+        }
         else
             res.status(200).json(q_res.rows[0]);
     });
 });
 
+/*
+ * GET /api/events/countEvents
+ * Get number of all events reserved in given time span.
+ *
+ * Request Parameters:
+ *  query:
+ *    date_from <string | Date> - if string, must be of form 'YYYY-MM-DD'
+ *    date_to   <string | Date> - if string, must be of form 'YYYY-MM-DD'
+ *    status    <string> default 'approved' - one of 'approved', 'denied', 'pending', 'all'
+ * Response:
+ *  200: successfully retrieved
+ *    <object>
+ *      count <int>
+ *  500: other postgres error
+ */
+router.get('/countEvents', (req, res) => {
+    pool.query(queries.getAllEventsCount, [ req.query.date_from, req.query.date_to, req.query.status ], (q_err, q_res) => {
+        if (q_err){
+            res.status(500).json({ err: 'PSQL Error: ' + q_err.message });
+            console.log(q_err.message);
+        }
+        else
+            res.status(200).json(q_res.rows[0]);
+    });
+});
 
 /*
  * GET /api/events
  * List events (summary) within date range, in ascending order by date/time.
- * 
- *                  SAMPLE
- * let query = {
- *      params: {
- *          date_from:"2020-08-09",
- *          date_to:"2020-08-09",
- *      }
- * }
- * axios.get("backend.com/api/events", query)
- *
  *
  * Request Parameters:
  *  query:
@@ -74,8 +93,6 @@ router.get('/countTickets', (req, res) => {
  *      time_start    <Date>
  *      hosts         <array[object]> - if type = 'detailed'
  *        uid         <int>
- *        email       <string>
- *        phone       <string>
  *        first_name  <string>
  *        last_name   <string>
  *        img_profile <string>
@@ -147,8 +164,15 @@ router.get('', (req, res) => {
  *      time_start    <Date>
  *      hosts         <array[object]>
  *        uid         <int>
- *        email       <string>
- *        phone       <string>
+ *        first_name  <string>
+ *        last_name   <string>
+ *        img_profile <string>
+ *        bio         <string>
+ *        school      <string>
+ *        major       <string>
+ *        grad_year   <string>
+ *      attendees     <array[object]>
+ *        uid         <int>
  *        first_name  <string>
  *        last_name   <string>
  *        img_profile <string>
@@ -194,7 +218,7 @@ router.get('/:eid', (req, res) => {
  *  201: successfully created
  *  403: attempting to create event for non-self user
  *  406: required field missing
- *  409: thumbnail already in use
+ *  409: thumbnail already in use or time unavailable
  *  500: other postgres error
  */
 router.post('', verifyFirebaseIdToken, async (req, res) => {
@@ -212,6 +236,35 @@ router.post('', verifyFirebaseIdToken, async (req, res) => {
         host_school = req.body.hosts[0].school;
     }
 
+    // convert to Date object
+    const time_start = new Date(req.body.time_start);
+
+    // make Zoom event
+    const zoom_res = await zoom.createMeeting(req.body.title, time_start);
+    if (!zoom_res) {
+        res.status(500).json({ err: 'Zoom API failed to create a new meeting' });
+        return;
+    }
+
+    // make Google Calendar event
+    const time_end = new Date(time_start);
+    time_end.setHours(time_end.getHours() + 1);
+    const gcal_id = await gcal.createGcalEvent(
+        req.body.title,
+        host_name,
+        req.profile.email,
+        req.body.description,
+        zoom_res.url,
+        zoom_res.id,
+        time_start,
+        time_end
+    );
+
+    if (!gcal_id) {
+        res.status(500).json({ err: 'Google Calendar API failed to create a new event' });
+        return;
+    }
+
     const values = [
         host_name,
         host_school,
@@ -220,27 +273,45 @@ router.post('', verifyFirebaseIdToken, async (req, res) => {
         req.body.description,
         req.body.requirements,
         req.body.thumbnail_id,
-        '', // zoom link empty for now
-        '', // zoom id empty for now
-        req.body.time_start,
+        zoom_res.url, 
+        zoom_res.id,
+        gcal_id,
+        time_start,
         'pending', // default status pending        
     ];
 
+    // insert into PSQL
     const client = await pool.connect();
     try {
         await client.query('BEGIN');
-        const valid_thumb = (await client.query(thumbnail_queries.checkThumbnail, [ req.body.thumbnail_id ])).rows.length > 0;
+
+        // check thumbnail not in use
+        const valid_thumb = (await client.query(
+            thumbnail_queries.checkThumbnailAvailable,
+            [ req.body.thumbnail_id ]
+        )).rows[0].available;
+
+        // check time still available
+        time_end.setHours(time_start.getHours() + 2);
+
+        const valid_time = (await client.query(
+            queries.checkTimeAvailable,
+            [ time_start, time_end ]
+        )).rows[0].available;
+
+        // if valid, add this event
         if (!valid_thumb) {
             await client.query('COMMIT');
             res.status(409).json({ err: 'Thumbnail already in use' });
+        } else if (!valid_time) {
+            await client.query('COMMIT');
+            res.status(409).json({ err: 'Time unavailable' });
         } else {
             await client.query(thumbnail_queries.setThumbnailUsed, [ req.body.thumbnail_id ]);
 
             const eid = (await client.query(queries.createEvent, values)).rows[0].eid;	
             for (const host of req.body.hosts)
                 await client.query(queries.createHost, [ host.uid, eid ]);
-
-            await client.query('COMMIT');
 
             // send email
             emails.sendEventSubmittedEmail(
@@ -249,8 +320,235 @@ router.post('', verifyFirebaseIdToken, async (req, res) => {
                 req.body.title
             );
 
+            await client.query('COMMIT');
             res.status(201).send();
+        }
 
+    // if not valid, delete the Zoom and GCal events
+    } catch (err) {
+        await client.query('ROLLBACK');
+        if (err.code === '23502') // not_null_violation
+            res.status(406).json({ err: 'Required field missing' });
+        else
+            res.status(500).json({ err: 'PSQL Error: ' + err.message });
+
+        if (!(await zoom.deleteMeeting(zoom_res.id)))
+            console.log('Failed to delete Zoom meeting', zoom_res.id);
+        if (!(await gcal.deleteGcalEvent(gcal_id)))
+            console.log('Failed to delete Gcal event', gcal_id);
+    } finally {
+        client.release();
+    }
+});
+
+/*
+ * PATCH /api/events/{eid}
+ * Update an event.
+ *
+ * Authorization:
+ *  Firebase ID Token
+ *
+ * Request Parameters
+ *  path:
+ *    eid <int> required
+ *
+ * Request Body:
+ *  <object>
+ *    host_name              <string>
+ *    host_school            <string>
+ *    host_bio               <string>
+ *    title                  <string>
+ *    description            <string>
+ *    requirements           <string>
+ *    thumbnail_id           <int> - id of the new thumbnail, if changing
+ *    time_start             <Date>
+ *    status                 <string> - either or 'approved', 'denied'
+ *
+ * Response:
+ *  201: successfully updated
+ *  403: must be an admin to update an event
+ *  406: eid missing
+ *  409: thumbnail already in use
+ *  500: other postgres error
+ */
+router.patch('/:eid', verifyIsAdmin, async (req, res) => {
+    if (!req.params.eid) {
+        res.status(406).json({ err: 'eid is required' });
+    	return;
+    }
+
+    if (req.body.status && req.body.status !== 'approved' && req.body.status !== 'denied') {
+        res.status(406).json({ err: 'status must be "approved" or "denied"' });
+        return;
+    }
+
+    const values = [
+        req.body.host_name|| null,
+        req.body.host_school || null,
+        req.body.host_bio || null,
+        req.body.title || null,
+        req.body.description || null,
+        req.body.requirements || null,
+        req.body.thumbnail_id || null,
+        req.body.time_start || null,
+        req.body.status || null,
+        req.params.eid
+    ];
+
+    const client = await pool.connect();
+    try {
+        await client.query('BEGIN');
+
+        // if thumbnail got updated, check thumbnail not in use
+        let valid_thumb = true;
+        if (req.body.thumbnail_id) {
+            valid_thumb = (await client.query(
+                thumbnail_queries.checkThumbnailAvailable,
+                [ req.body.thumbnail_id ]
+            )).rows[0].available;
+        }
+
+        // if start time got updated, check time still available
+        let valid_time = true;
+        if (req.body.time_start) {
+            const time_start = new Date(req.body.time_start);
+            const time_end = new Date(time_start);
+            time_end.setHours(time_start.getHours() + 2);
+
+            valid_time = (await client.query(
+                queries.checkTimeAvailable,
+                [ time_start, time_end ]
+            )).rows[0].available;
+        }
+
+        // if either thumbnail or time is invalid, return 409
+        if (!valid_thumb) {
+            await client.query('COMMIT');
+            res.status(409).json({ err: 'Thumbnail already in use' });
+        } else if (!valid_time) {
+            await client.query('COMMIT');
+            res.status(409).json({ err: 'Time unavailable' });
+
+        // otherwise, update the event
+        } else {
+
+            // get and keep track of event pre-update
+            const orig_event = (await client.query(queries.getEvent, [ req.params.eid ])).rows[0];
+
+            // update thumbnail in PSQL if it has changed
+            if (req.body.thumbnail_id) {
+                await client.query(thumbnail_queries.replaceThumbnail, [
+                    orig_event.img_thumbnail,
+                    req.body.thumbnail_id,
+                ]);
+            }
+
+            // update the event proper
+            await client.query(queries.updateEvent, values);
+
+            // send email and possibly update Gcal/Zoom if an approval took place
+            if (orig_event.status !== 'approved' && req.body.status === 'approved') {
+                
+                // get final event infos
+                const event_time = req.body.time_start || orig_event.time_start;
+                const event_title = req.body.title || orig_event.title;
+
+                // if necessary, update Zoom
+                if (req.body.time_start || req.body.title) {
+                    if (!(await zoom.updateMeeting(orig_event.zoom_id, event_title, event_time))) {
+                        await client.query('ROLLBACK');
+                        client.release();
+                        res.status(500).json({
+                            err: 'Zoom API failed to update meeting ' + orig_event.zoom_id 
+                        });
+                        return;
+                    }
+                }
+
+                // if necessary, update Gcal
+                if (req.body.time_start || req.body.title || req.body.host_name || req.body.description) {
+                    const gcal_event_patch = {}
+                    if (req.body.time_start) {
+                        const time_start = new Date(req.body.time_start);
+                        const time_end = new Date(time_start);
+                        time_end.setHours(time_start.getHours() + 1);
+
+                        gcal_event_patch.start = {
+                            dateTime: time_start.toISOString(),
+                            timeZone: TIMEZONE
+                        };
+
+                        gcal_event_patch.end = {
+                            dateTime: time_end.toISOString(),
+                            timeZone: TIMEZONE
+                        };
+                    }
+
+                    if (req.body.title)
+                        gcal_event_patch.summary = req.body.title;
+
+                    if (req.body.host_name || req.body.description) {
+                        const event_host = req.body.host_name || orig_event.host_name;
+                        const event_desc = req.body.description || orig_event.description;
+                        gcal_event_patch.description = `
+                            <html>
+                              <p>
+                                A Schefs event hosted by ${event_host}.
+                                <br><br>
+                                ${event_desc}
+                                <br><br>
+                                Meeting Link: <a href=${orig_event.zoom_link}>${orig_event.zoom_link}</a><br>
+                                Meeting ID: ${orig_event.zoom_id}
+                              </p>
+                            </html>
+                        `;
+                    }
+                    
+                    if (!(await gcal.updateGcalEvent(orig_event.gcal_id, gcal_event_patch))) {
+                        await client.query('ROLLBACK');
+                        client.release();
+                        res.status(500).json({
+                            err: 'Google Calendar API failed to update the event ' + orig_event.gcal_id
+                        });
+                        return;
+                    }
+                }
+
+                // send approval email
+                for (const host of orig_event.hosts) {
+                    emails.sendEventApprovedEmail(
+                        host.email,
+                        host.first_name,
+                        event_title,
+                        moment.tz(event_time, 'America/New_York').format('dddd, MMMM D, YYYY'),
+                        moment.tz(event_time, 'America/New_York').format('h:mm A z'),
+                        process.env.BASE_URL + '/' + orig_event.eid,
+                        req.body.zoom_link || orig_event.zoom_link
+                    );
+                }
+
+            // or send email and delete Zoom/Gcal if a denial took place
+            } else if (orig_event.status !== 'denied' && req.body.status === 'denied') {
+                for (const host of orig_event.hosts) {
+                    emails.sendEventDeniedEmail(
+                        host.email,
+                        host.first_name,
+                        req.body.title || orig_event.title,
+                        req.body.description || orig_event.description,
+                        req.body.requirements || orig_event.requirements,
+                        req.body.host_bio || orig_event.host_bio
+                    );
+                }
+
+                if (!(await zoom.deleteMeeting(zoom_res.id)))
+                    console.log('Failed to delete Zoom meeting', zoom_res.id);
+                if (!(await gcal.deleteGcalEvent(gcal_id)))
+                    console.log('Failed to delete Gcal event', gcal_id);
+            }
+
+            // if all went well, commit
+            await client.query('COMMIT');
+            res.status(201).send();
         }
     } catch (err) {
         await client.query('ROLLBACK');
@@ -267,8 +565,6 @@ router.post('', verifyFirebaseIdToken, async (req, res) => {
  * POST /api/events/{eid}/tickets
  * Reserve ticket associated with a specified event and user.
  *
- * TODO: Restrict to self/admin
- *
  * Authorization:
  *  Firebase ID Token
  *
@@ -278,11 +574,17 @@ router.post('', verifyFirebaseIdToken, async (req, res) => {
  *
  * Response:
  *  201: success
+ *  403: must be self or admin
  *  404: event and/or user do not exist
  *  406: event sold out or ticket already reserved
  *  500: other postgres error
  */
 router.post('/:eid/tickets', verifyFirebaseIdToken, async (req, res) => {
+    if (parseInt(req.profile.uid) !== parseInt(req.params.uid) && !req.profile.is_admin) {
+        res.status(403).send();
+        return;
+    }
+
     const client = await pool.connect();
     try {
         await client.query('BEGIN');
@@ -297,7 +599,6 @@ router.post('/:eid/tickets', verifyFirebaseIdToken, async (req, res) => {
             ];
 
             await client.query(queries.reserveTicket, values);
-            await client.query('COMMIT');
 
             // handle emails
             
@@ -340,6 +641,7 @@ router.post('/:eid/tickets', verifyFirebaseIdToken, async (req, res) => {
                 event.title
             ));
 
+            await client.query('COMMIT');
             res.status(201).send();
         }
     } catch (err) {
@@ -359,8 +661,6 @@ router.post('/:eid/tickets', verifyFirebaseIdToken, async (req, res) => {
  * DELETE /api/events/{eid}/tickets/{uid}
  * Delete ticket associated with a specified event and user.
  *
- * TODO: Restrict to self/admin
- *
  * Authorization:
  *  Firebase ID Token
  *
@@ -371,10 +671,16 @@ router.post('/:eid/tickets', verifyFirebaseIdToken, async (req, res) => {
  *
  * Response:
  *  204: success
+ *  403: must be self or admin to delete a ticket
  *  404: event and/or user and/or ticket do not exist
  *  500: other postgres error
  */
 router.delete('/:eid/tickets/:uid', verifyFirebaseIdToken, (req, res) => {
+    if (parseInt(req.profile.uid) !== parseInt(req.params.uid) && !req.profile.is_admin) {
+        res.status(403).send();
+        return;
+    }
+
     const values = [ req.params.eid, req.params.uid ];
     pool.query(queries.deleteTicket, values, (q_err, q_res) => {
         if (q_err)
@@ -428,7 +734,7 @@ router.get('/:eid/countTickets', (req, res) => {
  *    <boolean> 
  *  500: other postgres error
  */
-router.get('/:eid/:uid/ticketstatus', (req, res) => {
+router.get('/:eid/:uid/ticketstatus', verifyIsAdmin, (req, res) => {
     pool.query(queries.checkTicketStatus, [ req.params.eid, req.params.uid ], (q_err, q_res) => {
         if (q_err){
             res.status(500).json({ err: 'PSQL Error: ' + q_err.message });
@@ -447,8 +753,7 @@ router.get('/:eid/:uid/ticketstatus', (req, res) => {
 
 /*
  * GET /api/events/{eid}/tickets
- * Get array of ticket objects for event of specified eid containing 
- * user's first name, last name, and uid.
+ * Get array of ticket objects for event of specified eid.
  *
  * Request Parameters:
  *  path:
@@ -463,8 +768,7 @@ router.get('/:eid/:uid/ticketstatus', (req, res) => {
  *  404: event does not exist
  *  500: other postgres error
  */
-router.get('/:eid/tickets', (req, res) => {
-    // check auth and other stuff here
+router.get('/:eid/tickets', verifyIsAdmin, (req, res) => {
     pool.query(queries.getReservedTickets, [ req.params.eid ], (q_err, q_res) => {
         if (q_err)
             res.status(500).json({ err: 'PSQL Error: ' + q_err.message });
@@ -476,8 +780,7 @@ router.get('/:eid/tickets', (req, res) => {
 
 /*
  * GET /api/events/{eid}/comments
- * Get array of comment objects for event of specified eid containing 
- * cid, uid, name, body, time_created
+ * Get array of comment objects for event of specified eid.
  *
  * Request Parameters:
  *  path:
@@ -497,7 +800,6 @@ router.get('/:eid/tickets', (req, res) => {
  *  500: other postgres error
  */
 router.get('/:eid/comments', (req, res) => {
-    // check auth and other stuff here
     pool.query(queries.getComments, [ req.params.eid ], (q_err, q_res) => {
         if (q_err)
             res.status(500).json({ err: 'PSQL Error: ' + q_err.message });
@@ -527,8 +829,7 @@ router.get('/:eid/comments', (req, res) => {
  *  500: other postgres error
  */
 
-router.post('/:eid/comment', (req, res) => {
-    // check auth and other stuff here
+router.post('/:eid/comment', verifyFirebaseIdToken, (req, res) => {
     const values = [req.body.user_id, req.body.name, req.body.body, req.body.school, req.params.eid]
     pool.query(queries.postComment, values, (q_err, q_res) => {
         if (q_err){
@@ -538,7 +839,6 @@ router.post('/:eid/comment', (req, res) => {
             res.status(201).send()
         }
     })
-
 });
 
 module.exports = router;
