@@ -2,7 +2,7 @@ const defaultTheme = require('tailwindcss/defaultTheme')
 
 const fontFamily = defaultTheme.fontFamily;
 fontFamily['sans'] = [
-  'Roboto', // <-- Roboto is a default sans font now
+  'arial', // <-- Roboto is a default sans font now
   'system-ui'// <-- you may provide more font fallbacks here
 ];
 
